refactor: move Genre domain model to genre package
