--- src/main/java/ru/jerael/booktracker/backend/domain/model/Genre.java
@@ -1,3 +1,0 @@
-package ru.jerael.booktracker.backend.domain.model;
-
-public record Genre(Integer id, String name) {}
