+++ src/main/java/ru/jerael/booktracker/backend/domain/model/genre/Genre.java
@@ -1,0 +1,3 @@
+package ru.jerael.booktracker.backend.domain.model.genre;
+
+public record Genre(Integer id, String name) {}
