+++ src/main/java/ru/jerael/booktracker/backend/domain/usecase/genre/GetGenreByIdUseCase.java
@@ -1,6 +1,6 @@
 package ru.jerael.booktracker.backend.domain.usecase.genre;
 
-import ru.jerael.booktracker.backend.domain.model.Genre;
+import ru.jerael.booktracker.backend.domain.model.genre.Genre;
 
 public interface GetGenreByIdUseCase {
     Genre execute(Integer id);
