+++ src/main/java/ru/jerael/booktracker/backend/domain/usecase/genre/GetGenresUseCase.java
@@ -1,6 +1,6 @@
 package ru.jerael.booktracker.backend.domain.usecase.genre;
 
-import ru.jerael.booktracker.backend.domain.model.Genre;
+import ru.jerael.booktracker.backend.domain.model.genre.Genre;
 import java.util.Set;
 
 public interface GetGenresUseCase {
