+++ src/main/java/ru/jerael/booktracker/backend/web/mapper/GenreWebMapper.java
@@ -1,6 +1,6 @@
 package ru.jerael.booktracker.backend.web.mapper;
 
-import ru.jerael.booktracker.backend.domain.model.Genre;
+import ru.jerael.booktracker.backend.domain.model.genre.Genre;
 import ru.jerael.booktracker.backend.web.annotation.WebMapper;
 import ru.jerael.booktracker.backend.web.dto.genre.GenreResponse;
 import java.util.LinkedHashSet;
