+++ src/test/kotlin/ru/jerael/booktracker/backend/application/usecase/book/UpdateBookDetailsUseCaseImplTest.kt
@@ -13,7 +13,7 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 import ru.jerael.booktracker.backend.domain.exception.NotFoundException
 import ru.jerael.booktracker.backend.domain.exception.UnprocessableContentException
-import ru.jerael.booktracker.backend.domain.model.Genre
+import ru.jerael.booktracker.backend.domain.model.genre.Genre
 import ru.jerael.booktracker.backend.domain.model.author.Author
 import ru.jerael.booktracker.backend.domain.model.book.Book
 import ru.jerael.booktracker.backend.domain.model.book.BookStatus
