+++ src/test/kotlin/ru/jerael/booktracker/backend/application/usecase/genre/GetGenresUseCaseImplTest.kt
@@ -7,7 +7,7 @@ import io.mockk.junit5.MockKExtension
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
-import ru.jerael.booktracker.backend.domain.model.Genre
+import ru.jerael.booktracker.backend.domain.model.genre.Genre
 import ru.jerael.booktracker.backend.domain.repository.GenreRepository
 
 @ExtendWith(MockKExtension::class)
