+++ src/test/kotlin/ru/jerael/booktracker/backend/factory/book/BookDomainFactory.kt
@@ -1,6 +1,6 @@
 package ru.jerael.booktracker.backend.factory.book
 
-import ru.jerael.booktracker.backend.domain.model.Genre
+import ru.jerael.booktracker.backend.domain.model.genre.Genre
 import ru.jerael.booktracker.backend.domain.model.author.Author
 import ru.jerael.booktracker.backend.domain.model.book.*
 import ru.jerael.booktracker.backend.domain.model.language.Language
